backend/requirements.txt
fastapi
uvicorn[standard]
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic